+++ server/src/controllers/handlers/handlersCreateProducts.js
@@ -1,31 +1,25 @@
 
-// const createProduct = require('../Products/createProduct')
+const {createProduct,calculateAverageRating} = require('../Products/createProduct')
 
-// const calculateAverageRating = (ratingArray)=> {
-//     console.log('RATINGARRAY',ratingArray);
-//     const validRatings = ratingArray.filter(rating => typeof rating === 'number' && !isNaN(rating));
-//     const totalRating = validRatings.reduce((sum, rating) => sum + rating, 0);
-//     return validRatings.length > 0 ? totalRating / validRatings.length : 0;
-//   }
 
-// const handlersCreateProducts = async (req, res)=>{
+const handlersCreateProducts = async (req, res)=>{
     
-//     try{
-//         const data = req.body;
-//         const newProduct =await createProduct(data);
+    try{
+        const data = req.body;
+        const newProduct =await createProduct(data);
         
-//         const averageRating = calculateAverageRating(newProduct.rating);
-//         newProduct.averageRating = parseFloat(averageRating.toFixed(2));
+        const averageRating = calculateAverageRating(newProduct.rating);
+        newProduct.averageRating = parseFloat(averageRating.toFixed(2));
         
-//         res.statuts(400).json(newProduct);
+        res.status(200).json(newProduct);
 
-//     }catch{
-//         console.log("error crate", "req", req)
-
-
-
-//     }
-// }
+    }catch(err){
+        console.log("error crate", err.message)
 
 
-// module.exports = handlersCreateProducts;
+
+    }
+}
+
+
+module.exports = handlersCreateProducts;
